♻️ [컬럼 경량화]타입 파일 분리
